refactor: package structure and add environment profile support
Updated package names from `capstone` to `me.sallim` and renamed root project to `api`. Added Spring profile support and `.env` integration in `application.yml`. Introduced new dependencies, such as Actuator and WebSocket, and changed test controller to a health check endpoint for the `dev` profile.
