+++ .example.env
@@ -1,3 +1,6 @@
+# cp .example.env .env
+
 DB_URL=db_url
 DB_USERNAME=db_username
 DB_PASSWORD=db_password
+SPRING_PROFILES_ACTIVE=dev
