+++ src/test/java/me/sallim/api/ServerApplicationTests.java
@@ -1,4 +1,4 @@
-package capstone.Server;
+package me.sallim.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
